Add migration to drop obsolete kombu tables and daemon timestamps
The changes in the daemon to make it event based, have gotten rid
of Celery, the daemon task timestamps and the concept of the daemon
user. This commit adds the migrations for both backends to delete
the corresponding tables and entries in the DbSetting table from
the database. Since this is a lossy migration, there is no downgrade
of the migration possible and, by extension, available