Implement the LazyChoice param type for click-based verdi commands
The click based verdi commands use decorators to define the options
and arguments of the command. These are executed at import time.
In general we attempt to put all the heavy imports within the body
of the commands, to keep verdi and its autocomplete fast and responsive.
However, sometimes those options require data that needs a heavy import.
An example is the choices for the process state of the `verdi work list`
command, which require the import of ProcessState from plumpy, which
was slowing everything down.

The LazyChoice allows to pass a callable instead of a list, such that
the import will only be called at runtime when the command is ran.