Define the 'options' of JobProcess as a PortNamespace
Originally the `options` for the JobProcess was specified as a single
non database storable `InputPort` that took a nested dictionary.
Because of this solution, the internal structure was not autocompleted
in the `ProcessBuilder`. By redefining the `options` port as a full blown
`PortNamespace` and each member within it as its own `Port`, a help string
can be defined and in the shell this can be displayed through the process
builder. Additionally, this allows the input validation to be done through
the same framework of plumpy as all the other inputs.

Since this was the last use of voluptuous, the requirement of this package
has been removed from the setup requirements