Fix bug in the copy method of Sealable nodes
The copy method of the base Node class was copying over all the
attributes, except for the sealed key. Subsequently, the copy
implementation of the Sealable mixin was deleting the updatable
attributes from the clone, by iterating over the attributes of
the original node. This would trigger an exception, because even
though the original node has the sealed attribute, it was excluded
in the Node.copy() implementation and so calling _del_attr on it
would raise. By iterating over the attributes of the clone node, the
problem is fixed.

Note however, that the base node implementation should really have
no knowledge whatsoever of the Sealable mixin and its sealed key.
However, removing it now leads to other problems, because by copying
it there, we can no longer delete the attributes afterwards, because
it will hit the ModificationNotAllowed exception, due to the sealed
attribute being present.