Fix minor bug in delete_nodes utility due to typo
Added a test that would have caught this bug. For this it was
necessary to set the verbosity to 2, because the bug was only
triggered in the branches of code for high verbosity. Additionally,
the tests created a simple graph of base Node instances, which
means that the logic involving Calculation nodes and CALL links
was not hit. The new test explicitly creates Calculation nodes with
CALL links to test this case