Implement a generic IdentifierParamType for verdi commands
A very common pattern for any verdi command will be an option or an argument
that takes on or multiple values that is supposed to identify a specific orm
entity. In AiiDA there are three types of identifier:

 * ID
 * UUID
 * STRING

Most entities have both an ID as well as a UUID that identifies them. Some
have an additional string based identifier, such as a label or a name.

Here we define the class IdentifierParamType, which can be used to create
reusable arguments and options that will load ORM instances based on the
passed values, by deducing whether they are meant to repesent an ID, UUID
or STRING. Note that since the type is lost and all values will be strings,
one has to guess the type. The value will be attempted to be interpreted
as an ID, UUID and then STRING, in that order.

This approach can lead to ambiguities in some situations, where if an
intended (partial) UUID can be interpreted as a valid ID. Or when a valid
STRING also happens to be a valid ID. However, as long as the users do
not use labels that are valid integers and use the full UUID in the worst
case, chances for these clashes should be highly unlikely.