Add a mechanism to always solve identifier type ambiguity
All ORM entities can be identified with an ID, UUID or optionally a
STRING like identifier. Any potential ambiguity between an ID and UUID,
for example when a partial UUID is also a valid ID, can always be solved
by including at least one dash in the UUID. The ambiguity when a STRING
identifier is also a valid ID or UUID is a little more tricky. We add the
convention here that if the last character of the identifier is an
exclamation point, it will always be interpreted as a STRING type like
identifier. Consider three entities with the following ID, UUID and STRING:

    1  b6420d91-7577 'some_name'
    2  d9c4c081-f6bb '1'
    3  0cf82610-3ae8 'b6420d91-7577'

Trying to select either the second or third entity by their STRING identifier
will always result in the first entity being matched, given that their STRING
identifiers also match the ID and UUID, respectively of the first entity,
To break this one simply needs to use the identifier:

	IdentifierParam.convert('1!')
    IdentifierParam.convert('b6420d91-7577!')

The added exclamation mark will break the ambiguity by forcing the method to
interpret the identifier as a STRING type.

This behavior is implemented in the `convert` method of the IdentifierParam
parameter type and tests have been added that test the functionality.