Refactoring of IdentifierParamTypes and OrmEntityLoader classes
This was necessary to allow to refactor the older load_node and load_group
functions to use the same implementation as the OrmEntityLoader classes. These
changes allow load_node and load_group to leverage the OrmEntityLoader class to
do the heavy lifting while keeping the same function signature and functionality.
The old query builder utility in `aiida.orm.utils` that these functions where
relying on, which was reimplemented in a more general way in the OrmEntityLoader,
was removed. The BaseTranslator of the REST API was relying on this function but
has been reworked to be able to use the OrmEntityLoader as well.