Allow base orm class to be overridden for OrmEntityLoader
The subclasses of the OrmEntityLoader have to define a base orm class
for which any entity that is to be matched and loaded has to be a
sub class of. However, sometimes a user may want to narrow the query
set even more and make the requirement on the base class even more
specific. To allow this, the get_query_builder and load_entity methods
now have the optional sub_classes argument, which takes a tuple of ORM
classes, each of which has to be a sub class of the base orm class.

When valid, the query will be limited to entities which are a sub class
of these specific orm sub classes.

This now also allows the IdentifierParamType and its subclasses to, upon
construction, define a set of entry points, whose classes will be passed
as this sub_classes tuple when a value needs to be converted. The reason
to taking entry point strings and not the actual ORM classes is to prevent
to have to load the database environment upon import of the verdi commands
in order to keep the autocompletion and interface quick and snappy.