Upgrade SqlAlchemy dependency requirement to 1.2.12
To make the code base compatible with the new version, expressions produced
by the querybuilder had to include casts from JSONB. The cast that needs to
happen depends on the type of the value, but the order of the type check
and the cast cannot be guaranteed in the old syntax. Therefore we
replace the attribute queries from chains to cases, such that the type
is checked before it is casted.