Interactive Option, multiline input, PluginParameterType, Code setup (#1611)
## Missing (suggested to create separate issue & pull request)
* Tests for `aiida.control.code.CodeBuilder`

## Added

### ConditionalOption
`aiida.cmdline.optiontypes.conditional.py`
A `click.Option` subclass, which is required only if a callback function returns True. This callback function recieves the command's context, which means the condition may include values of other parameters.

Limitations:
 * does not work with `is_eager=True`
 * options with callbacks (and `is_eager=False` get executed last, but do not rely on a particular order between them. -> conditions should not depend on other options with callbacks

### InteractiveOption
`aiida.cmdline.optiontypes.interactive.py`
A `ConditionalOption` subclass, which
* can prompt for missing values (not given as option on CLI)
* can give a list of viable options (depends on parameter type) when "?" is entered at the prompt
* can depend on other options (like `ConditionalOption`)
* prompting for all `ConditionalOptions` of a command can be disabled by passing a `--non-interactive` flag

### multiline input
`aiida.cmdline.utils.multi_line_input.py`
Functions for opening an editor for the user to enter multi line text as input in the style of `git commit`.

Available functionality:
* `edit_pre_post`: open editor to create or change Prepend / Append bash scripts, combined, optionally displaying the configuration as a comment
* `edit_comment`: open editor to attach comments to nodes, or change existing comments

### PluginParameterType
A `click.ParameterType` for plugin names, optionally validates that the plugin loader can find the entry point name in the given category. Can print a list of available plugins on interactive prompt, tab-completion ready.

### CodeBuilder
`aiida.control.code`
A small utility class to simplify and standardize creating a code using the builder pattern. It is in the `control` module because it is primarily useful when working interactively or writing a tool.

### Reusable options
A number of reusable options: `--non-interactive`, `--input-plugin` and others

## Modified

* verdi code setup translated to click, using all of the above, introducing non-interactive mode
* verdi code setup serves as a usage example for `InteractiveOption`, `PluginParameterType` and `edit_pre_post()`