Reorganization of cmdline params and commands
The aiida.cmdline.params.arguments have been reorganized in similar
fashion to aiida.cmdline.params.options. The module contains multiple
files with variants of arguments and options, respectively. Any
predefined arguments and options, using the Overridable class, are then
defined in the __init__.py such that they are also automatically
exposed in that level of the module.