Migrate `verdi rehash` to use the click infrastructure
Also changed the interface options of the command. The --class-name has
been replaced by -e/--entry-point, since that is a more robust way of
defining a class in AiiDA. Additionally, the --all option has been
removed, since that would be implied if no identifier arguments would
be specified but only an entry point.