Prevent continuing a task that has already terminated
The situation may occur where a worker picks up a continuation task
and the process actually reaches a terminal state, however, the worker
may fail to communicate this to RabbitMQ. In that case, the task will
be sent again by RabbitMQ and be picked up by another worker. This
worker will try to load it but the checkpoint will already have been
deleted by the previous worker after the task terminated.

We override the _continue method of the ProcessLauncher to check whether
the node corresponding to the task is already marked as terminated,
before even attempting to load the process from the checkpoint. When
the node is actually in a terminal state, the corresponding future is
recreated based on its actual terminal state. For FINISHED we return the
outputs, for KILLED we set a KilledError exception and for EXCEPTED we
return a PastException with the exception that had been set before.

For this final part to work, we also had to implement that any exception
that triggered the EXCEPTED state would also be saved for the node as
an attribute, such that it could be retrieved later.