Generalize the PluginParamType to support multiple entry point groups
Often one wants to define a command line parameter whose values should
be valid entry points, but not necessarily restricted to one particular
entry point group. Here we update PluginParamType to support definition
of a tuple of valid entry point groups.

This change does bring a complication with it. Since entry point names
are only guaranteed to be unique within an entry point group, but not
necessarily across groups, the situation may arise where a parameter
is defined with support for one or more entry point groups that share
an entry point with a common name. The conversion and autocompletion
of entry points with a minimal format, i.e. only the entry point name
and without the entry point group, may be ambiguous. In this case, the
conversion should fail and auto-completion should always return the
valid entry point strings in FULL format. In cases where there is no
risk of ambiguity, the PluginParamType should support the user to pass
all three types of entry point string format and adapt the auto-complete
to the form of the incomplete value as much as possible.