Implement the Exit exception to allow setting finish status for workfunctions (#1631)
There was no method yet to exit from a workfunction and specify an exit code
that should be used to set the finish status of the corresponding node.
However, this is an important feature to have. Here we choose to implement
a specific exception, aiida.work.exceptions.Exit, which will be treated
differently from all other exceptions by the workflow engine. When this is
raised, as opposed to marking the workfunction as EXCEPTED, the exit code
is retrieved from the exception class and is set as the finish status on the
node of the workfunction.