Overload PortNamespace mutable properties upon exposing
When exposing the ports of a process one would expect that the
receiving port namespace would not just inherit the ports but
also the mutable properties of the PortNamespace, such as the
default, valid_type, help, validator and required attribute.

The method that was responsible for creating the new namespace
and exposing the ports, ProcessSpec._expose_ports, failed to do
this. However, plumpy already defined a method that did just
this, PortNamespace.absorb(). To fix this problem, we decided
to move the expose functionality to plumpy and leverage the
absorb method, which already properly overloaded mutable
properties of the donor namespace to the host namespace.

The only other addition is that the implementation of exposing
ports in plumpy did not yet have a memory, as required by the
implementation in aiida-core. This implementation, along with
the proper namespace property inheritance has now been released
with plumpy v0.10.3, which is the new dependency.

The expose methods in plumpy now also takes an optional dictionary
namespace_options which can be used to override even the properties
of the source namespace that would be inherited. If the ports are
exposed into a namespace that does not yet exist in the target
namespace, it will be constructed using this namespace_options
dictionary as constructor keyword arguments. Note that these
namespace options will *not* be used for any sub namespaces that
will have to be created along the way, nor will they be used to
override any of the properties of already existing sub namespaces.