Implement support for process input serialization (#1574)
Adds a `serializer` keyword argument to the `InputPort`, which can be used
to specify a serialization function that is called if the input is not an AiiDA data
type. This is implemented by adding a serialize method to the `InputPort` and
`PortNamespace`, which takes an input or input dictionary and applies the
serialization.