Converted verdi user to click yo
Added:
 * UserParamType
 * USER option
 * USER argument

Currently UserParamType only accepts the email.  We may want to extend
this to take an ID.

Made an extension to InteractiveOption that allows for a
contextual_default.  What this does is get the ctx from the click runner
which can then be used in the callback function to get the default.
This was used so that the user could specify a user email on the command
line and the properties (first name, last, etc) are automatically
filled in.

Patched up the tests, although they're somewhat basic.