Added prompt_fn option not to prompt interactive prompts
E.g. when a previous option decides if the option should be printed or not
Also fixed the tests.

NOTE: if the wrong number of interactive options is passed in a
test, this will enter an infinite loop and will hang the tests!

To decide if we want to have a way to prevent this (e.g. a max_prompt_count variable?)