Fixing the logic of the daemon that checks for the scheduler state
In 0.12, it was (correctly) avoiding to ask also for a list of job ids
if the scheduler supported querying by users. Now instead it was asking
BOTH for user and job ids, that is not supported in general by the
schedulers (see e.g. docstring of Scheduler._get_joblist_command).
This fixes this behavior.