Fixing the logic of the daemon that checks for the scheduler state (#1688)
In 0.12.0 it was (correctly) avoiding to ask also for a list of job ids
if the scheduler supported querying by users. Now instead it was asking
for both the user and job ids, which is not supported in general by the
schedulers (see e.g. docstring of Scheduler._get_joblist_command).