Ensure that WorkChain._do_step return None, int or ExitCode (#1704)
In the case where the return value of the stepper was not of type
None, int or ExitCode, yet the stepper considered itself to be finished,
for example when the last step was executed, the return value of the
stepper would be returned, which could have any type. However, the
result of _do_step should only be an exit status, i.e. int or ExitCode.
To fix this, we explicitly check the return value of the stepper and
convert any type other than int or ExitCode to None.