Add unique constraints on UUID column of Node database model
Even though the UUID of Nodes are supposed to be unique, there was no
explicit constraint on the column, allowing two nodes with identical
UUIDS to be stored without complaint from the database. However, these
nodes can now no longer be loaded through their UUID leaving the
database in an unusable state. This commit adds the migrations for both
backends and adds a test to check that attempting to store a Node with
a UUID that already exists will raise an IntegrityError.