Migrate `verdi user` to the click infrastructure (#1661)
Added the `UserParamType` and the corresponding `USER` argument and option.
Currently UserParamType only accepts the email. An extension to `InteractiveOption`
was made that allows for a `contextual_default`.  What this does is get the `ctx` from
the click runner which can then be used in the callback function to get the default.
This was used so that the user could specify a user email on the command line and
the properties (first name, last, etc) are automatically filled in.