commands/data: rename `list.py` to avoid type collision
Otherwise even a specific import like

    from aiida.cmdline.commands.data.list import _list

will then overload the type `list` to be the module list instead of the
type. So, expressions like `list([])` in the same scope (or below)
as the import will generate an error stating that a module is not callable.

Also rename the `export.py` for the same reason and to be consistent.

Furthermore, this way it is also clear that those files do not contain
standalone commands but rather helper functions.