verdi cmdline/commands improvements (#1755)
* commands/data: delay load_dbenv further

* commands/data/bands: use `finally` to close files

* commands/data: add `plugins` subcommand

* commands/data: rename `list.py` to avoid type collision

Otherwise even a specific import like

    from aiida.cmdline.commands.data.list import _list

will then overload the type `list` to be the module list instead of the
type. So, expressions like `list([])` in the same scope (or below)
as the import will generate an error stating that a module is not callable.

Also rename the `export.py` for the same reason and to be consistent.

Furthermore, this way it is also clear that those files do not contain
standalone commands but rather helper functions.

* cmdline/utils/pluginable: fix list_commands

* commands/data: move sub-command import

We must ensure that the commands are defined at the point Pluginable's
`list_commands` is called, so they must be imported in global scope.
Since we delayed the `load_dbenv` further this is not a problem anymore.

* cmdline/baseclass: pass along to click in run()

For verdi subcommands this now runs the corresponding click command
instead of going through the list of available_subcommands. Via the
`Pluginable` class we therefore get proper plugin support until we can
drop the `VerdiSubcommand` class entirely.

* commands/data/cif: improve command description