Homogenize the interface of `verdi quicksetup` and `verdi setup`
These two commands should essentially provide the exact same interface
to the user, as the only difference is that `verdi quicksetup` will
try to guess most of the required variables to setup a profile.
However, the functionality once all required variables are defined
should be identical for the two commands.

 * Both commands now have the same options and arguments
 * All options and arguments are defined once as Overridable*
 * All options now have a help string