Enable auto completion of verdi parameter values
This feature is not supported by `click` out of the box, which
only auto completes the commands, sub commands and the parameters
themselves. For the parameter value completion, we add the package
click-completion, which is then initialized in the the module
`aiida.cmdline.commands`, just like the `verdi` command group.
This now allows `ParamType` instances to implement the `complete`
method, which takes the `ctx` and the `incomplete` string and
should return a list of strings, with the completed values.