Refactor imports to reduce loading time for `verdi` and keep it snappy
The recent move of `verdi` to the click infrastructure, also brought
a change to the `verdi computer configure` commands. Each transport
now has its own sub command within that group and these are loaded
dynamically at the bottom of`aiida.cmdline.commands.cmd_computer`.
To dynamically create these commands, the required parameters are
generated from the `_valid_auth_options` from the `Transport` class
implementation, which therefore needs to be imported. The `ssh` module
imported the `paramiko` library at the top level, which if one has
`gss_api` installed will cause the load time to rise to 0.8 seconds.
By moving these imports within the functions, the average load time
for just `verdi` is reduced to ~0.2 which is the maximum acceptable.