Check for `test_` keyword in repo path only on last directory
The `create_configuration` function would check the repository path for the
`test_` keyword and would raise if it was absent for a test profile or present
for a non-test profile. However, the check was performed on whether the prefix
was present in the entire path. This is too strict, as this would disallow any
path that even downstream would contain the string `test_`. Here, we change
the test to just check the last directory of the absolute path, which was the
intended behavior