Add `verdi work show`
In working on this issue, it was also found that the `verdi calculation`
commands, no longer accepted `InlineCalculation` nodes but only calcs of
the type `JobCalculation`. This was has been addressed as well.

Finally, with the `verdi` click group being defined in the module init
of `aiida.cmdline`, as well as all sub commands being imported, such that
they are registered, the sub commands having to import the verdi group
to attach themselves, caused a cyclic import. By moving the `verdi`
group to its own file, this cyclicity is broken.