Synchronize heuristics of `verdi work list` and `verdi calculation list`
The logic for both these commands should be that for default values of
all the options, all entries are shown that are considered 'active'.
The `-a/--all` flag will cause all entries to be shown, independent
of their 'activeness'. Finally, this list can be limited by passing the
`-p/--past-days` flag. This issue was indirectly addressed by the move
of `verdi` to the `click` infrastructure as now both commands use the
same options, defined in `aiida.cmdline.params.option`.
This commit just updates the help text of the `-a/--all` flag to make
sure that it correctly reflects its behavior.