Refactor ORM collection and collection entries to always take the backend
Certain ORM collection entries, such as the AuthInfo class, require an
instance of the backend class, to retrieve other backend collections
and or entries. To ensure that these get properly constructed the concept
is abstracted by the Collection and CollectionEntry class which will are
to be sub classed for the various ORM classes. These base classes will
ensure that the backend is passed at construction time and provides a
property to retrieve it.