Fix bug in the RemoteData._clean method
The `clean_remote` free function was not imported. Fix the import,
and add a test to trigger this code path. As a result, a bug was
found in the AuthInfo and User backend implementations, where the
wrong type for `backend` argument were being passed to the constructor
by the Collection.