Properly set future exception if transport fails to open
The `do_open` callback of the `request_transport` method of the
`TransportQueue` was not catching any exceptions when calling
`transport.open()`. As a result, the future's exception would
never be set, causing the thread to wait forever. We wrap the
open call in try except block and set the exception on the future
accordingly when the open call excepts