Include `transport.open()` call in exponential backoff retry coroutine (#1852)
Improved the transport tasks of `JobProcess` by wrapping the call for the
opening of the transport through `transport.open()` in the exponential backoff
retry coroutine as well. This was not the case and a failure to open the transport
itself would fail the transport task before it could call the actual function.