Test that TransportQueue respects the safe open interval (#1851)
There was a suspect case that the safe open interval for the transport
was not being respected by the `TransportQueue` because there were
debug statements that exceeded the usual `30` second interval for the
ssh transport. However, it turns out that in the current `develop` the
default interval for the `SshTransport` is hard coded to `5` seconds.
This commit just adds a test to check that the safe open interval is always
respected by the `TransportQueue`.