Abstract common code in `_setup_db_record` for various Processes (#1863)
The `_setup_db_record` method of the `Process` class was completely
reimplemented for the `JobProcess` sub class even though there is a
significant overlap of functionality, leading to discrepancies, such
as the `process_label` attribute not being set for `JobProcess` instances.

Abstract the common functionality such that `JobProcess` can call the
super implementation. The only real difference was the setting up of
inputs links which has been moved to `_setup_db_inputs` and is implemented
separately for the `JobProcess` sub class.