Fix plugin fixture setup profile (#1856)
Use `aiida.control.profiles.setup_profiles` in the fixture

Due to a refactoring issue, it was trying to use `aiida.cmdline.commands.cmd_setup.setup`
which is the click command. This was caused by the click migration and not caught because
the tests were skipped due to a bug in `pg_test` that was fixed in another pull request.