Determine active nodes in `verdi calculation list` on process state (#1873)
The base behavior of `verdi calculation list`, just as its counterpart
`verdi work list`, should be to list all the "active" entries. However,
what constitutes an "active" entry for these two differs. For the work
command, this is determined based on the process state of the nodes.
Any node that is not in a terminal process state, is considered active.
However, for legacy reasons, in `verdi calculation list` the activeness
was still being based on the calculation state, which was not always in
sync with the process state, leading to inconsistent displays.

Here we update the behavior of `verdi calculation list` to determine the
activeness of calculation nodes based on their process state, thus
making it directly equivalent with the behavior of `verdi work list` and
`verdi process list.