Make TransportTasks interruptable
All TransportTasks for the JobProcess Waiting state have been cast
into coroutines, for the purpose of allowing them to yield to other
functionality such as interruptions. However, once in the tasks itself
it could not be interrupted. This led to any interruption being issued
after a task was started, to only be effectuated after the task had
already completed.

Here we wrap all transport tasks in the `interruptable_task` utility
function, which wraps a transport task coroutine and turns it execution
into a future, while also defining a cancel flag that is passed as a
reference to the transport task coroutine. When the Waiting state is
now interrupted, this is piped through to the future, which sets the
cancelled flag. The transport task can now check this flag in its body
whenever it wants to, to check whether it has been cancelled, in which
case it should raise `plumpy.CancelledError`, which will then properly
bubble up to the step function.

The `exponential_backoff_retry` had to be updated to ignore this
particular exception and reraise it to the transport task that called
it.