Ensure that SqlAlchemy Node._del_attr throws AttributeError
The abstract implementation specifies that Node._del_attr should
throw an AttributeError if the specified key does not correspond
to an existing attribute. The SqlAlchemy implementation however,
threw a ValueError. This was not caught, because this behavior
was only tested for unstored nodes, which follows a different code
path. With the introduction of the Sealable concept, where certain
attributes can be modified even after storing the node, this new
code path became available, thus exposing the bug.