Use equality in OrmEntityLoader UUID filter for complete UUIDs
If the `_get_query_builder_uuid_identifier` method of `OrmEntityLoader`
gets an identifier that constitutes a complete UUID, the query builder
can use an equality operator instead of the like operator. This should
be preferred since the former is a lot more efficient database
operation. To verify whether the provided UUID is a complete UUID, we
simply try to construct a UUID object out of it, which should work only
for complete UUIDs and throw a ValueError in all other cases.