Have execmanager.kill_calculation not fail if job did not exist
The `execmanager.kill_calculation` function was calling the `kill`
method of the scheduler and was raising a `RemoteOperationError`
if the return value of that call was anything but True. However,
in some cases, the reason for the return value not being True, was
because the job no longer existed. In the case of the DirectScheduler
plugin, the kill command would print an error to stderr which would
be interpreted as an error and so the function returned False, causing
the exception to be raised. This would make it impossible to distinguish
between a kill call that actually failed to kill the job and one that
merely failed because the job did not exist. If the job no longer exists
the kill command can be interpreted as successful, as its inteded result
has already been achieved.

To ameliorate this, when `scheduler.kill` returns anything else but True,
we check with the scheduler if the job we were trying to kill actually
still exists or has already been killed or completed. If the latter is
the case, we consider the kill command as successfull and do not raise.