Factor out the JobCalculation._presubmit from submit_calculation
The `execmanager.submit_calculation` has recently been transformed into
a transport task of the JobProcess, given that it requires transport to
actually submit the job. Since opening the transport or the actual submit
operations can fail, the execmanager function is wrapped in the exponential
backoff retry coroutine. This means that the body of the execmanager function
will potentially be executed multiple times, which then requires that the
function is idempotent.

Originally, the `submit_calculation` would call both the `_presubmit` step
of the JobCalculation class, which would write the input files, store them
in the raw input folder in the repository and build up the calculation info
that the transport needed to actually copy over the input files and submit
the job. The first part does not involve the transport and so should not be
part of the transport task.

To separate these concerns, `execmanager.submit_calculation` is refactored
to separate the presubmit step from the actial transport submit. The former
has been moved to the `run` method of the `JobProcess`, which means that it
should only be executed once, as it should. The transport task now then only
includes the actual copying over of the input files and the submission of the
job to the scheduler.